+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Changed
+- Rename 'max' to 'limit'
+- Using a 'long' instead of a 'int' for amount and limit
+
+### Fixed
+- The amount resetting to 0 when the max value of the number type was reached
+
 ## [1.4.1]
+
 ### Fixed
 - Fixed editor dependencies